+++ src/base/index.js
@@ -39,6 +39,7 @@ export { ManagePollBuilder } from './operations/manage_poll_builder'
 export { ManageVoteBuilder } from './operations/manage_vote_builder'
 export { ManageAccountSpecificRuleBuilder } from './operations/manage_account_specific_rule_builder'
 export { RemoveAssetPairOpBuilder } from './operations/remove_asset_pair_op_builder'
+export { CreateKYCRecoveryRequestBuilder } from './operations/create_kyc_recovery_request_builder'
 
 export * from './strkey'