+++ CHANGELOG.md
@@ -8,18 +8,18 @@ Please check our [developers guide](https://gitlab.com/tokend/developers-guide)
 for further information about branching and tagging conventions.
 
 ## [Unreleased]
-#### Added
+- `kycRecovery` to `WalletsManager`
 - `encryptSecretSeed`,`decryptSecretSeed` and `deriveSessionKey` methods to crypto.js
 - `sessionId` and `sessionKey` arguments for constructor in Wallet class
 
 ## [1.9.0-x.0] - 2019-07-01
-#### Added
+### Added
 - `validateWithoutSnakeCased` flag for validators `validateCreatorDetails` method
 
 ## [1.8.0] - 2019-06-22
 
 ## [1.8.0-rc.1] - 2019-06-17
-#### Added
+### Added
 - `isEmptyBodyAllowed` flag for api-caller `_call` method
 
 ## [1.8.0-rc.0] - 2019-06-06