+++ CHANGELOG.md
@@ -8,6 +8,7 @@ Please check our [developers guide](https://gitlab.com/tokend/developers-guide)
 for further information about branching and tagging conventions.
 
 ## [Unreleased]
+### Added
 - `kycRecovery` to `WalletsManager`
 - `encryptSecretSeed`,`decryptSecretSeed` and `deriveSessionKey` methods to crypto.js
 - `sessionId` and `sessionKey` arguments for constructor in Wallet class